Move streamers to own package
